pyproject.toml
[project]
name = "coursework"
version = "0.1.0"
description = ""
authors = [
    {name = "lengvin",email = "zve565ro@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.4.2,<9.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
black = "^25.1.0"
mypy = "^1.18.2"
isort = "^6.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
line_length = 119
